fix: references to tiki_sdk_dart
also improves the example
