+++ example/pubspec.yaml
@@ -13,7 +13,7 @@ dependencies:
   http: ^1.0.0
   html: ^0.15.3
   uuid: ^3.0.7
-  tiki_sdk_dart: 
+  tiki_trail:
     path: ../.
 
 dev_dependencies:
