+++ test/cache/receipt/receipt_repository_test.dart
@@ -5,17 +5,17 @@
 
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
-import 'package:tiki_sdk_dart/cache/license/license_model.dart';
-import 'package:tiki_sdk_dart/cache/license/license_repository.dart';
-import 'package:tiki_sdk_dart/cache/license/license_use.dart';
-import 'package:tiki_sdk_dart/cache/license/license_usecase.dart';
-import 'package:tiki_sdk_dart/cache/payable/payable_model.dart';
-import 'package:tiki_sdk_dart/cache/payable/payable_repository.dart';
-import 'package:tiki_sdk_dart/cache/receipt/receipt_model.dart';
-import 'package:tiki_sdk_dart/cache/receipt/receipt_repository.dart';
-import 'package:tiki_sdk_dart/cache/title/title_model.dart';
-import 'package:tiki_sdk_dart/cache/title/title_repository.dart';
-import 'package:tiki_sdk_dart/utils/bytes.dart';
+import 'package:tiki_trail/cache/license/license_model.dart';
+import 'package:tiki_trail/cache/license/license_repository.dart';
+import 'package:tiki_trail/cache/license/license_use.dart';
+import 'package:tiki_trail/cache/license/license_usecase.dart';
+import 'package:tiki_trail/cache/payable/payable_model.dart';
+import 'package:tiki_trail/cache/payable/payable_repository.dart';
+import 'package:tiki_trail/cache/receipt/receipt_model.dart';
+import 'package:tiki_trail/cache/receipt/receipt_repository.dart';
+import 'package:tiki_trail/cache/title/title_model.dart';
+import 'package:tiki_trail/cache/title/title_repository.dart';
+import 'package:tiki_trail/utils/bytes.dart';
 import 'package:uuid/uuid.dart';
 
 void main() {
