+++ test/cache/title/title_repository_test.dart
@@ -5,9 +5,9 @@
 
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
-import 'package:tiki_sdk_dart/cache/title/title_model.dart';
-import 'package:tiki_sdk_dart/cache/title/title_repository.dart';
-import 'package:tiki_sdk_dart/utils/bytes.dart';
+import 'package:tiki_trail/cache/title/title_model.dart';
+import 'package:tiki_trail/cache/title/title_repository.dart';
+import 'package:tiki_trail/utils/bytes.dart';
 import 'package:uuid/uuid.dart';
 
 void main() {
