+++ test/l0/storage/storage_service_test.dart
@@ -9,10 +9,10 @@ import 'dart:typed_data';
 import 'package:nock/nock.dart';
 import 'package:pointycastle/api.dart';
 import 'package:test/test.dart';
-import 'package:tiki_sdk_dart/l0/auth/auth_service.dart';
-import 'package:tiki_sdk_dart/l0/storage/storage_service.dart';
-import 'package:tiki_sdk_dart/utils/rsa/rsa.dart';
-import 'package:tiki_sdk_dart/utils/rsa/rsa_private_key.dart';
+import 'package:tiki_trail/l0/auth/auth_service.dart';
+import 'package:tiki_trail/l0/storage/storage_service.dart';
+import 'package:tiki_trail/utils/rsa/rsa.dart';
+import 'package:tiki_trail/utils/rsa/rsa_private_key.dart';
 import 'package:uuid/uuid.dart';
 
 import '../auth/auth_nock.dart';
