+++ test/node/backup/backup_repository_test.dart
@@ -8,8 +8,8 @@ import 'dart:typed_data';
 
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
-import 'package:tiki_sdk_dart/node/backup/backup_model.dart';
-import 'package:tiki_sdk_dart/node/backup/backup_repository.dart';
+import 'package:tiki_trail/node/backup/backup_model.dart';
+import 'package:tiki_trail/node/backup/backup_repository.dart';
 import 'package:uuid/uuid.dart';
 
 void main() {
