+++ test/node/transaction/transaction_repository_test.dart
@@ -8,10 +8,10 @@ import 'dart:typed_data';
 
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
-import 'package:tiki_sdk_dart/node/block/block_model.dart';
-import 'package:tiki_sdk_dart/node/block/block_repository.dart';
-import 'package:tiki_sdk_dart/node/transaction/transaction_model.dart';
-import 'package:tiki_sdk_dart/node/transaction/transaction_repository.dart';
+import 'package:tiki_trail/node/block/block_model.dart';
+import 'package:tiki_trail/node/block/block_repository.dart';
+import 'package:tiki_trail/node/transaction/transaction_model.dart';
+import 'package:tiki_trail/node/transaction/transaction_repository.dart';
 import 'package:uuid/uuid.dart';
 
 void main() {
