+++ test/node/xchain/xchain_repository_test.dart
@@ -9,9 +9,9 @@ import 'dart:typed_data';
 import 'package:sqlite3/common.dart';
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
-import 'package:tiki_sdk_dart/node/xchain/xchain_model.dart';
-import 'package:tiki_sdk_dart/node/xchain/xchain_repository.dart';
-import 'package:tiki_sdk_dart/utils/bytes.dart';
+import 'package:tiki_trail/node/xchain/xchain_model.dart';
+import 'package:tiki_trail/node/xchain/xchain_repository.dart';
+import 'package:tiki_trail/utils/bytes.dart';
 import 'package:uuid/uuid.dart';
 
 void main() {
