+++ test/utils/bytes_test.dart
@@ -1,9 +1,14 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
 import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:test/test.dart';
-import 'package:tiki_sdk_dart/utils/bytes.dart';
-import 'package:tiki_sdk_dart/utils/compact_size.dart';
+import 'package:tiki_trail/utils/bytes.dart';
+import 'package:tiki_trail/utils/compact_size.dart';
 
 void main() {
   group('Bytes Tests', () {
