+++ test/utils/merkel_tree_test.dart
@@ -9,11 +9,11 @@ import 'dart:typed_data';
 import 'package:pointycastle/export.dart';
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
-import 'package:tiki_sdk_dart/node/key/key_model.dart';
-import 'package:tiki_sdk_dart/node/key/key_service.dart';
-import 'package:tiki_sdk_dart/node/transaction/transaction_model.dart';
-import 'package:tiki_sdk_dart/node/transaction/transaction_service.dart';
-import 'package:tiki_sdk_dart/utils/merkel_tree.dart';
+import 'package:tiki_trail/node/key/key_model.dart';
+import 'package:tiki_trail/node/key/key_service.dart';
+import 'package:tiki_trail/node/transaction/transaction_model.dart';
+import 'package:tiki_trail/node/transaction/transaction_service.dart';
+import 'package:tiki_trail/utils/merkel_tree.dart';
 import 'package:uuid/uuid.dart';
 
 import '../in_mem.dart';
