+++ test/utils/rsa_test.dart
@@ -8,9 +8,9 @@ import 'dart:typed_data';
 
 import 'package:pointycastle/api.dart';
 import 'package:test/test.dart';
-import 'package:tiki_sdk_dart/utils/rsa/rsa.dart';
-import 'package:tiki_sdk_dart/utils/rsa/rsa_private_key.dart';
-import 'package:tiki_sdk_dart/utils/rsa/rsa_public_key.dart';
+import 'package:tiki_trail/utils/rsa/rsa.dart';
+import 'package:tiki_trail/utils/rsa/rsa_private_key.dart';
+import 'package:tiki_trail/utils/rsa/rsa_public_key.dart';
 
 void main() {
   group('RSA Tests', () {
